requirements.txt

pandas
requests
ta
websocket-client
python-dotenv
telepot
